`WishboneMaster` and `RecordProcessor` done
Refactored these modules to the new design. Wrote some documentation.
Wrote Hedgehog tests for both of these and fixed bugs.

The tests for `recordProcessorC` are complete. The test for
`wishboneMasterT` is probably not, but it is more efficient to continue
with the other modules and test the whole system in the end.